chore(code): delete unused code
